requirements.txt
Django==4.2.7
djangorestframework==3.14.0
environs==9.5.0
pre_commit==3.5.0
flake8==6.1.0
black==23.11.0
autoflake==2.2.1
pydrive==1.3.1